Extract LogicalScheduler passes 0-2 into standalone functions
Move place_lrs, assign_vgpr_tiles, and place_grs out of the
LogicalScheduler class into separate files under a new
Tensile/Components/Subtile/passes/ package.

Each pass is now a standalone function taking explicit typed parameters
(SchedulerConfig + schedule) instead of operating on self. This enables
independent unit testing and removes hidden state coupling.

Structural changes:
- New _types.py: all dataclasses, constants (TENSOR_SIDE), and helpers
  (fmt_mt) extracted from LogicalScheduler.py to break circular imports.
  LogicalScheduler.py re-exports everything for backward compatibility.
- New passes/__init__.py: re-exports place_lrs, assign_vgpr_tiles,
  VgprAllocation, place_grs, partition_tile_range.
- New passes/_utils.py: partition_tile_range(config, pi).
- New passes/place_lrs.py: place_lrs(config) -> LogicalSchedule.
- New passes/assign_vgpr_tiles.py: assign_vgpr_tiles(config, schedule)
  -> VgprAllocation dataclass (replaces setting self attributes).
- New passes/place_grs.py: place_grs(config, schedule) with helpers
  _build_gr_list, _build_lr_conflict_map, _has_lr_conflict,
  _distribute_grs.

LogicalScheduler.build() now calls the standalone functions and stores
VgprAllocation fields (unroll_factor, needs_unrolling, tile_peaks) on
self. Old methods removed from the class.

Tests updated to call standalone functions via the passes package.
All 60 tests pass.
