+++ projects/hipblaslt/tensilelite/Tensile/Components/Subtile/LogicalScheduler.py
@@ -35,317 +35,29 @@ from .ScheduleTypes import (
     EmittedSchedule,
     LogicalSchedule,
 )
-
-
-TENSOR_SIDE = {'A': 'A', 'B': 'B', 'SA': 'A', 'SB': 'B'}
-
-def fmt_mt(mt: int) -> str:
-    """Format MT iteration integer as display string: 0 → 'n', 1 → 'n+1', 2 → 'n+2'."""
-    return "n" if mt == 0 else f"n+{mt}"
-
-# ── Core primitives ─────────────────────────────────────────
-
-@dataclass
-class MFMATileRange:
-    """A rectangular range of MFMA tile coordinates for one read."""
-    subIterK_start: int
-    subIterK_end: int          # exclusive
-    tileId_start: int
-    tileId_end: int            # exclusive
-
-    @property
-    def subIterK_list(self) -> List[int]:
-        return list(range(self.subIterK_start, self.subIterK_end))
-
-    @property
-    def tileId_list(self) -> List[int]:
-        return list(range(self.tileId_start, self.tileId_end))
-
-    def fmt_k(self) -> str:
-        ids = self.subIterK_list
-        if len(ids) == 1:
-            return f"[{ids[0]}]"
-        return f"[{ids[0]},{ids[-1]}]"
-
-    def fmt_tiles(self) -> str:
-        return f"[{self.tileId_start}-{self.tileId_end - 1}]"
-
-
-# ── Config ──────────────────────────────────────────────────
-
-@dataclass
-class ReadGranularity:
-    """Load granularity for one operation on one tensor, measured in MFMA tiles.
-
-    mn: how many MFMA tiles in the M (for A/SA) or N (for B/SB) dimension
-    k:  how many subIterK steps one read covers
-    """
-    mn: int
-    k: int
-
-
-@dataclass
-class SchedulerConfig:
-    """Configuration for the MFMATile-based scheduler."""
-    numMFMATilesM: int    # MFMA tiles in M dimension (for A)
-    numMFMATilesN: int    # MFMA tiles in N dimension (for B)
-    numSubIterK: int      # subIterK steps within the macrotile
-    lrA: ReadGranularity
-    lrB: ReadGranularity
-    grA: ReadGranularity
-    grB: ReadGranularity
-    lrSA: Optional[ReadGranularity] = None
-    lrSB: Optional[ReadGranularity] = None
-    grSA: Optional[ReadGranularity] = None
-    grSB: Optional[ReadGranularity] = None
-    numPartitionsM: int = 1   # partition grid in M dimension
-    numPartitionsN: int = 1   # partition grid in N dimension
-
-    @property
-    def hasScale(self) -> bool:
-        return self.lrSA is not None and self.lrSB is not None
-
-    @property
-    def numPartitions(self) -> int:
-        return self.numPartitionsM * self.numPartitionsN
-
-    @property
-    def partitionSizeM(self) -> int:
-        assert self.numMFMATilesM % self.numPartitionsM == 0
-        return self.numMFMATilesM // self.numPartitionsM
-
-    @property
-    def partitionSizeN(self) -> int:
-        assert self.numMFMATilesN % self.numPartitionsN == 0
-        return self.numMFMATilesN // self.numPartitionsN
-
-    @staticmethod
-    def get_partition_candidates(tileInfoA, tileInfoB) -> list:
-        """Return partition candidates as [(numPartitionsM, numPartitionsN), ...].
-
-        Enumerates all divisors of MAX(M, N) in ascending order and
-        partitions the larger dimension. Starts with (1, 1).
-        This will only produces 1xN or Nx1 partitions to allow VGPR pressure reduction.
-        """
-        M = tileInfoA.localMMATileGrid[0]
-        N = tileInfoB.localMMATileGrid[0]
-        maxDim = max(M, N)
-
-        divisors = sorted(d for d in range(1, maxDim + 1) if maxDim % d == 0)
-
-        candidates = []
-        for d in divisors:
-            if N >= M:
-                candidates.append((1, d))
-            else:
-                candidates.append((d, 1))
-
-        return candidates
-
-
-
-# ── Schedule operation types ────────────────────────────────
-
-@dataclass
-class Emittable:
-    """Base for anything placed in an EmittedModule."""
-    kind: str = field(init=False, default="")
-
-
-@dataclass
-class MFMAPlacement(Emittable):
-    """MFMA operation consuming data for one subIterK."""
-    subIterK: int
-    tileA: MFMATileRange       # A tiles consumed
-    tileB: MFMATileRange       # B tiles consumed
-    deps: List['Dep'] = field(default_factory=list)      # populated by annotate_deps()
-    preOps: List['BaseOp'] = field(default_factory=list)     # populated by remove_cross_deps()
-    vgpr_tile_maps: Dict[str, List[dict]] = field(default_factory=dict)  # {tensor: [{groupIdx: vgprTileId}]} per unroll iter
-
-    def __post_init__(self):
-        self.kind = 'mfma'
-
-    def __str__(self):
-        return (f"MFMAs (MT n, subIterK {self.subIterK}  ) "
-                f"A : {self.tileA.fmt_tiles()} , B : {self.tileB.fmt_tiles()}")
-
-
-@dataclass
-class LRPlacement(Emittable):
-    """Local Read placement for one tensor in one subIterK slot."""
-    tensor: str                # 'A', 'B', 'SA', 'SB'
-    mtIteration: int           # 0 = current MT, 1 = next MT
-    tiles: MFMATileRange
-    subIterK_slot: int         # which subIterK this LR is placed in
-    partition: int = 0         # which partition this LR belongs to
-    deps: List['Dep'] = field(default_factory=list)      # populated by annotate_deps()
-    preOps: List['BaseOp'] = field(default_factory=list)     # populated by remove_cross_deps()
-    vgpr_tile_map: List[dict] = field(default_factory=list)  # [{tileId: vgprTileId}] per unroll iter
-
-    def __post_init__(self):
-        self.kind = 'lr'
-
-    def __str__(self):
-        return (f"LR {self.tensor.ljust(2)} (MT {fmt_mt(self.mtIteration)}, "
-                f"subIterK {self.tiles.fmt_k()}) {self.tiles.fmt_tiles()}")
-
-
-@dataclass
-class GRPlacement(Emittable):
-    """Global Read placement for one tensor in one subIterK slot."""
-    tensor: str                # 'A', 'B', 'SA', 'SB'
-    mtIteration: int           # 1 = next MT, 2 = two MTs ahead
-    tiles: MFMATileRange
-    subIterK_slot: int         # which subIterK this GR is placed in
-    partition: int = 0         # which partition this GR belongs to
-    deps: List['Dep'] = field(default_factory=list)      # populated by annotate_deps()
-    preOps: List['BaseOp'] = field(default_factory=list)     # populated by remove_cross_deps()
-
-    def __post_init__(self):
-        self.kind = 'gr'
-
-    def __str__(self):
-        return (f"GR {self.tensor} (MT {fmt_mt(self.mtIteration)}, "
-                f"subIterK {self.tiles.fmt_k()}) ids {self.tiles.fmt_tiles()}")
-
-
-# ── Per-subIterK container ──────────────────────────────────
-
-@dataclass
-class SubIterKSlot:
-    """All operations placed in one subIterK step."""
-    subIterK: int
-    mfma: Optional[MFMAPlacement] = None
-    lrs: List[LRPlacement] = field(default_factory=list)
-    grs: List[GRPlacement] = field(default_factory=list)
-
-
-# ── Dependency types ────────────────────────────────────────
-
-@dataclass
-class WaitGRCounts:
-    """Per-tensor inflight load counts for wait_gr preOp."""
-    A: int = 0
-    B: int = 0
-    SA: int = 0
-    SB: int = 0
-
-    def __str__(self):
-        parts = []
-        for t in ('A', 'B', 'SA', 'SB'):
-            v = getattr(self, t)
-            if v:
-                parts.append(f"{t}={v}")
-        return ",".join(parts) if parts else "0"
-
-
-@dataclass
-class BaseOp(Emittable):
-    """Base class for typed dependency operations in a before-chain."""
-
-    def __str__(self):
-        return self.kind
-
-
-@dataclass
-class WaitGROp(BaseOp):
-    """Wait for global reads to complete. Optionally includes a sync barrier."""
-    wait_gr_counts: Optional[WaitGRCounts] = None
-    has_sync: bool = False
-
-    def __post_init__(self):
-        self.kind = 'wait_gr'
-
-    def __str__(self):
-        if self.wait_gr_counts:
-            return f"{self.kind}({self.wait_gr_counts})"
-        return self.kind
-
-
-@dataclass
-class WaitLROp(BaseOp):
-    """Wait for local reads to complete. Optionally includes a sync barrier."""
-    has_sync: bool = False
-
-    def __post_init__(self):
-        self.kind = 'wait_lr'
-
-
-@dataclass
-class SyncOp(BaseOp):
-    """Standalone sync barrier."""
-    def __post_init__(self):
-        self.kind = 'sync'
-
-
-@dataclass
-class LRIncOp(BaseOp):
-    """LDS buffer swap for local reads on a specific tensor."""
-    tensor: str = ""
-
-    def __post_init__(self):
-        self.kind = 'lr_inc'
-
-    def __str__(self):
-        return f"lr_inc({self.tensor})"
-
-
-@dataclass
-class GRIncOp(BaseOp):
-    """Pointer update + LDS swap for global reads on a specific tensor."""
-    tensor: str = ""
-
-    def __post_init__(self):
-        self.kind = 'gr_inc'
-
-    def __str__(self):
-        return f"gr_inc({self.tensor})"
-
-
-@dataclass
-class SkipOp(BaseOp):
-    """Skip guard: compare LoopCounter and branch."""
-    compare: str = ""
-    value: int = 0
-    target: str = ""
-
-    def __post_init__(self):
-        self.kind = 'skip'
-
-    @property
-    def tensor(self) -> str:
-        return f"{self.compare}:{self.value}:{self.target}"
-
-    def __str__(self):
-        return f"skip({self.tensor})"
-
-
-@dataclass
-class Dep:
-    """Dependency on another placement (annotate_deps output)."""
-    ref: Union[LRPlacement, GRPlacement]
-    mt_offset: int = 0  # 0 = same MT, -1 = prev MT, -2 = two MTs back, ...
-
-
-
-
-# ── Emitted output ─────────────────────────────────────────
-
-@dataclass
-class EmittedModule:
-    """One emitted module with before-link for instruction scheduling.
-
-    Compatible with SubtileBasedInstructionScheduler.instructionSchedule().
-    Instructions are left empty at the logical level — filled during emission.
-    """
-    moduleId: int = -1
-    instructions: list = field(default_factory=list)
-    before: Optional[int] = None   # moduleId that must complete before this module
-    source: Optional[Emittable] = None
-
-    @property
-    def opType(self) -> str:
-        return self.source.kind if self.source else ""
+from ._types import (  # noqa: F401 — re-exported for backward compatibility
+    TENSOR_SIDE,
+    fmt_mt,
+    MFMATileRange,
+    ReadGranularity,
+    SchedulerConfig,
+    Emittable,
+    MFMAPlacement,
+    LRPlacement,
+    GRPlacement,
+    SubIterKSlot,
+    WaitGRCounts,
+    BaseOp,
+    WaitGROp,
+    WaitLROp,
+    SyncOp,
+    LRIncOp,
+    GRIncOp,
+    SkipOp,
+    Dep,
+    EmittedModule,
+)
+from .passes import place_lrs, assign_vgpr_tiles, place_grs, partition_tile_range
 
 
 # ── Main scheduler class ───────────────────────────────────
@@ -359,7 +71,6 @@ class LogicalScheduler:
 
     def __init__(self, config: SchedulerConfig):
         self.config = config
-        self.tensors: List[str] = ['A', 'B'] + (['SA', 'SB'] if config.hasScale else [])
         self._emitted: Optional[EmittedSchedule] = None
         self._preloop_emitted: Optional[EmittedSchedule] = None
         self._ngll_emitted: Optional[EmittedSchedule] = None
@@ -368,552 +79,8 @@ class LogicalScheduler:
     # ── Place LRs ─────────────────────────────────────────
 
     def _partition_tile_range(self, pi: int) -> dict:
-        """Return {'A': (start, end), 'B': (start, end)} for partition pi.
-
-        Uses COLUMN_MAJOR ordering: M (A) varies fastest, N (B) varies slowest.
-        """
-        cfg = self.config
-        # COLUMN_MAJOR: M is inner (pi % M), N is outer (pi // M)
-        piM = pi % cfg.numPartitionsM
-        piN = pi // cfg.numPartitionsM
-        a0 = piM * cfg.partitionSizeM
-        b0 = piN * cfg.partitionSizeN
-        return {'A': (a0, a0 + cfg.partitionSizeM),
-                'B': (b0, b0 + cfg.partitionSizeN)}
-
-    def place_LRs(self) -> LogicalSchedule:
-        """Place MFMAs and LRs based on read granularities.
-
-        Returns a list of partitions, each containing a list of SubIterKSlots.
-
-        Each LR prefetches data for the next subIterK group. Within-partition
-        prefetches use current partition tiles; cross-partition prefetches
-        (wrapping) use next partition tiles.
-
-        Two tracking mechanisms:
-        - loaded_ranges: tracks tile ranges in VGPR per side. Wrapping LRs
-          are only placed when the next partition's tiles aren't already loaded.
-        - placed: tracks (tensor, k-range, tile-range) of non-wrapping LRs
-          placed so far across partitions. Skips redundant K-prefetch when
-          the same data was already loaded by an earlier partition.
-        """
-        cfg = self.config
-        numP = cfg.numPartitions
-        part_ranges = [self._partition_tile_range(pi) for pi in range(numP)]
-
-        # Track which tile ranges are currently loaded in VGPR (for wrapping decisions).
-        loaded_ranges = {'A': {part_ranges[0]['A']},
-                         'B': {part_ranges[0]['B']}}
-
-        # Track placed K-prefetch LRs across partitions (for dedup).
-        placed = set()
-
-        partitions = []
-        for pi in range(numP):
-            cur, nxt = part_ranges[pi], part_ranges[(pi + 1) % numP]
-            is_last = (pi == numP - 1)
-
-            load = {}
-            for side in ('A', 'B'):
-                load[side] = is_last or nxt[side] not in loaded_ranges[side]
-
-            slots = self._place_LRs_for_partition(cur, nxt, is_last, load, placed)
-            for slot in slots:
-                for lr in slot.lrs:
-                    lr.partition = pi
-            partitions.append(slots)
-
-            for side in ('A', 'B'):
-                if load[side]:
-                    loaded_ranges[side] = {cur[side], nxt[side]}
-
-        return partitions
-
-    def _place_LRs_for_partition(self, cur: tuple, nxt: tuple,
-                                  is_last: bool,
-                                  load: dict,
-                                  placed: set) -> List[SubIterKSlot]:
-        """Place MFMAs and LRs for one partition."""
-        cfg = self.config
-        numK = cfg.numSubIterK
-        multi_part = cfg.numPartitions > 1
-
-        slots = [SubIterKSlot(subIterK=k) for k in range(numK)]
-        slot_mt = {}  # slot_k → lr_mt string, for MT-homogeneity enforcement
-
-        # MFMAs
-        for k in range(numK):
-            slots[k].mfma = MFMAPlacement(
-                subIterK=k,
-                tileA=MFMATileRange(k, k + 1, cur['A'][0], cur['A'][1]),
-                tileB=MFMATileRange(k, k + 1, cur['B'][0], cur['B'][1]),
-            )
-
-        # All tensors that can participate.
-        all_tensors = [('A', cfg.lrA), ('B', cfg.lrB)]
-        if cfg.hasScale:
-            all_tensors.append(('SA', cfg.lrSA))
-            all_tensors.append(('SB', cfg.lrSB))
-
-        # Place LRs grouped by k_gran.
-        # - Non-wrapping (K-prefetch): all tensors, deduped by placed set.
-        # - Wrapping (cross-partition): only tensors whose side needs loading.
-        for k_gran in sorted(set(g.k for _, g in all_tensors)):
-            group_all = [(t, g) for t, g in all_tensors if g.k == k_gran]
-            num_chunks = numK // k_gran
-            for chunk_idx in range(num_chunks):
-                next_chunk = (chunk_idx + 1) % num_chunks
-                is_wrap = (next_chunk == 0)
-                lr_mt = 1 if is_last and is_wrap else 0
-                lr_k_start = next_chunk * k_gran
-                lr_k_end = lr_k_start + k_gran
-                base_slot = chunk_idx * k_gran
-
-                # For wrapping chunks, only include tensors whose side is
-                # loading so that slot assignment reflects active tensors.
-                # A and B always participate (their wrapping is gated inside
-                # the loop) to keep slot indices stable for their k_gran group.
-                if is_wrap and multi_part:
-                    group = [(t, g) for t, g in group_all
-                             if t in ('A', 'B') or load['A' if t in ('A', 'SA') else 'B']]
-                else:
-                    group = group_all
-
-                # Group by side (A/SA together, B/SB together) for slot assignment
-                sides = [[(t, g) for t, g in group if t in ('A', 'SA')],
-                         [(t, g) for t, g in group if t in ('B', 'SB')]]
-                sides = [s for s in sides if s]
-
-                for side_idx, side in enumerate(sides):
-                    slot_k = base_slot + (side_idx % k_gran)
-                    # Redirect LRs away from slots committed to a different MT,
-                    # keeping each slot MT-homogeneous.
-                    # This reduce the number of wait_gr_sync needed as all LRs 
-                    # in the same subIterK wait for the same MT iterration.
-                    committed = slot_mt.get(slot_k)
-                    if committed is not None and committed != lr_mt:
-                        slot_k = numK - 1
-
-                    for tensor, gran in side:
-                        tile_range = nxt if (is_wrap or not multi_part) else cur
-                        side_key = 'A' if tensor in ('A', 'SA') else 'B'
-                        ts, te = tile_range[side_key]
-
-                        # Wrapping: use load dict. Non-wrapping: use placed set.
-                        if is_wrap and multi_part:
-                            if not load[side_key]:
-                                continue
-                        else:
-                            lr_key = (tensor, lr_k_start, lr_k_end, ts, te)
-                            if lr_key in placed:
-                                continue
-                            placed.add(lr_key)
-
-                        lr = LRPlacement(
-                            tensor=tensor,
-                            mtIteration=lr_mt,
-                            tiles=MFMATileRange(lr_k_start, lr_k_end, ts, te),
-                            subIterK_slot=slot_k,
-                        )
-                        slots[slot_k].lrs.append(lr)
-                        slot_mt[slot_k] = lr_mt
-
-        return slots
-
-    # ── Assign VGPR tile IDs (free-list allocation) ──────
-
-    def assign_vgpr_tiles(self, schedule: LogicalSchedule) -> LogicalSchedule:
-        """Assign physical vgprTileIds to all placements (A, B, SA, SB).
-
-        Free-list allocator with per-tensor FIFO queues, iterated until
-        convergence (or max 4 unroll iterations).
-
-        Three phases:
-          1. Scan all MFMAs to find last read position for each
-             (tensor, tileId, k_data_group) key.
-          2. Walk execution order in a loop: each iteration feeds the
-             previous next_iter as the starting active state.  Appends
-             one tile-map dict per iteration to each placement's list.
-             Stops when next_iter matches the seeded state (convergence).
-          3. Record unroll_factor, needs_unrolling, and max tile_peaks.
-
-        Keys use a unified formula parameterized by ReadGranularity:
-          key = (tensor, (tileId // lr_gran.mn) * lr_gran.mn, (k // lr_gran.k) * lr_gran.k)
-
-        Sets self.tile_peaks (per-tensor max across unrolls),
-        self.needs_unrolling, self.unroll_factor.
-        """
-        cfg = self.config
-        numK = cfg.numSubIterK
-        MAX_UNROLL = 8
-
-        lr_grans = {'A': cfg.lrA, 'B': cfg.lrB}
-        if cfg.hasScale:
-            lr_grans['SA'] = cfg.lrSA
-            lr_grans['SB'] = cfg.lrSB
-
-        # ── Phase 1: find last MFMA read for each key ──
-        last_read = {}  # key -> flat position
-        for pi, slots in enumerate(schedule):
-            for slot in slots:
-                if not slot.mfma:
-                    continue
-                pos = pi * numK + slot.subIterK
-                k = slot.subIterK
-                for tensor in self.tensors:
-                    side = TENSOR_SIDE[tensor]
-                    tileRange = slot.mfma.tileA if side == 'A' else slot.mfma.tileB
-                    gran = lr_grans[tensor]
-                    for t in tileRange.tileId_list:
-                        group = (t // gran.mn) * gran.mn
-                        k_chunk = (k // gran.k) * gran.k
-                        last_read[(tensor, group, k_chunk)] = pos
-
-        # ── Phase 2: iterate until convergence ──
-        from collections import deque
-
-        class _FreeList:
-            __slots__ = ('free', 'next_id', 'active_count', 'peak')
-            def __init__(self):
-                self.free = deque()
-                self.next_id = 0
-                self.active_count = 0
-                self.peak = 0
-            def alloc(self):
-                if self.free:
-                    vid = self.free.popleft()  # FIFO for convergence
-                else:
-                    vid = self.next_id
-                    self.next_id += 1
-                self.active_count += 1
-                self.peak = max(self.peak, self.active_count)
-                return vid
-            def release(self, vid):
-                self.free.append(vid)
-                self.active_count -= 1
-
-        max_peaks = {t: 0 for t in self.tensors}
-        carry_active = {}
-        all_next_iters = []     # next_iter from each iteration, for cycle detection
-
-        pools = {t: _FreeList() for t in self.tensors}
-
-        for unroll_iter in range(MAX_UNROLL):
-            if unroll_iter == 0:
-                active = {}
-            else:
-                active = dict(carry_active)
-                # Reset active_count to match carry_active (tiles that survived
-                # as live from the previous iteration's wrapping LRs).
-                for t in self.tensors:
-                    pools[t].active_count = sum(
-                        1 for key in active if key[0] == t)
-
-            next_iter = {}
-
-            for pi, slots in enumerate(schedule):
-                for slot in slots:
-                    pos = pi * numK + slot.subIterK
-                    k = slot.subIterK
-
-                    # ── MFMA reads: look up or seed ──
-                    if slot.mfma:
-                        for tensor in self.tensors:
-                            side = TENSOR_SIDE[tensor]
-                            tileRange = slot.mfma.tileA if side == 'A' else slot.mfma.tileB
-                            gran = lr_grans[tensor]
-                            tile_map = {}
-                            for t in tileRange.tileId_list:
-                                group = (t // gran.mn) * gran.mn
-                                k_chunk = (k // gran.k) * gran.k
-                                key = (tensor, group, k_chunk)
-                                if key not in active:
-                                    active[key] = pools[tensor].alloc()
-                                tile_map[group] = active[key]
-                            slot.mfma.vgpr_tile_maps.setdefault(tensor, []).append(tile_map)
-
-                    # ── LR writes: allocate new tiles ──
-                    for lr in slot.lrs:
-                        tensor = lr.tensor
-                        is_wrapping = lr.mtIteration != 0
-                        target = next_iter if is_wrapping else active
-
-                        gran = lr_grans[tensor]
-                        tile_map = {}
-                        seen_keys = set()
-                        for t in lr.tiles.tileId_list:
-                            group = (t // gran.mn) * gran.mn
-                            for lk in lr.tiles.subIterK_list:
-                                k_chunk = (lk // gran.k) * gran.k
-                                key = (tensor, group, k_chunk)
-                                if key in seen_keys:
-                                    continue
-                                seen_keys.add(key)
-                                if key in target:
-                                    pools[tensor].release(target[key])
-                                vid = pools[tensor].alloc()
-                                target[key] = vid
-                                tile_map[group] = vid
-                        lr.vgpr_tile_map.append(tile_map)
-
-                    # ── Release tiles whose last read was at this position ──
-                    to_release = [key for key, lr_pos in last_read.items()
-                                  if lr_pos == pos and key in active]
-                    for key in to_release:
-                        pools[key[0]].release(active[key])
-                        del active[key]
-
-            # Track max peaks across iterations
-            for t in self.tensors:
-                max_peaks[t] = max(max_peaks[t], pools[t].peak)
-
-            # Check convergence: if this iteration's next_iter matches
-            # any previous iteration's next_iter, we found a cycle.
-            # The cycle period is (current_iter - matching_iter).
-            # All iterations from matching_iter to current_iter-1 form
-            # the repeating pattern; iterations before that are prologue.
-            converged = False
-            for prev_idx, prev_ni in enumerate(all_next_iters):
-                if next_iter == prev_ni:
-                    # Strip tile maps from the redundant convergence iteration.
-                    for pi2, slots2 in enumerate(schedule):
-                        for slot2 in slots2:
-                            if slot2.mfma:
-                                for tensor in self.tensors:
-                                    if tensor in slot2.mfma.vgpr_tile_maps:
-                                        slot2.mfma.vgpr_tile_maps[tensor].pop()
-                            for lr2 in slot2.lrs:
-                                lr2.vgpr_tile_map.pop()
-                    converged = True
-                    break
-            if converged:
-                break
-
-            # Carry next_iter forward as active for next iteration
-            all_next_iters.append(next_iter)
-            carry_active = next_iter
-        else:
-            assert False, (f"assign_vgpr_tiles did not converge after "
-                           f"{MAX_UNROLL} unroll iterations")
-
-        # ── Phase 3: record results ──
-        # unroll_factor = number of unique iterations (convergence iteration excluded)
-        self.unroll_factor = unroll_iter
-        self.needs_unrolling = self.unroll_factor > 1
-        self.tile_peaks = max_peaks
-        return schedule
-
-    # ── Place GRs ─────────────────────────────────────────
-
-    def _build_gr_list(self, schedule: LogicalSchedule, part_ranges, offsetMT, offsetPartition,
-                             debug=False):
-        """Phase 1: Build ordered GR list from placed MFMAs.
-
-        For each partition × subIterK, derive target partition/MT from
-        the MFMA and offsets. Add GRs (A, B, SA, SB) with tile and K
-        ranges snapped to GR granularity. Dedup within same MT level,
-        then remove n+1 entries that also appear at n+2 (cross-MT dedup).
-        
-        For each subIterK, we apply offsetMT on MT and offsetPartition on partition.
-
-        Returns list of (tensor, mt_str, tile_start, tile_end,
-                         k_start, k_end, gr_gran).
-        """
-        cfg = self.config
-        numP = cfg.numPartitions
-
-        seen = set()
-        gr_list = []
-
-        for pi in range(numP):
-            partition_slots = schedule[pi]
-
-            target_pi = (pi + offsetPartition) % numP
-            wraps = (pi + offsetPartition) >= numP
-            mt_val = offsetMT + (1 if wraps else 0)
-
-            target_range = part_ranges[target_pi]
-
-            for slot in partition_slots:
-                k = slot.mfma.subIterK
-
-                items = [('A', target_range['A'], cfg.grA),
-                         ('B', target_range['B'], cfg.grB)]
-                if cfg.hasScale:
-                    items.append(('SA', target_range['A'], cfg.grSA))
-                    items.append(('SB', target_range['B'], cfg.grSB))
-
-                for tensor, (t_start, t_end), gr_gran in items:
-                    mn = gr_gran.mn
-                    k_gran = gr_gran.k
-
-                    gr_tile_start = (t_start // mn) * mn
-                    gr_tile_end = ((t_end + mn - 1) // mn) * mn
-
-                    gr_k_start = (k // k_gran) * k_gran
-                    gr_k_end = gr_k_start + k_gran
-
-                    key = (tensor, mt_val, gr_tile_start, gr_tile_end,
-                           gr_k_start, gr_k_end)
-                    if key in seen:
-                        continue
-                    seen.add(key)
-                    gr_list.append((tensor, mt_val, gr_tile_start,
-                                    gr_tile_end, gr_k_start, gr_k_end,
-                                    gr_gran))
-
-        # Cross-MT dedup: if a tile/k range appears at both n+1 and n+2,
-        # the n+1 load is redundant — the previous iteration's n+2 already
-        # wrote the same data into LDS.  Remove the n+1 duplicate.
-        base_mt = offsetMT
-        n2_keys = {(t, ts, te, ks, ke)
-                   for t, mt, ts, te, ks, ke, _ in gr_list
-                   if mt != base_mt}
-        gr_list = [entry for entry in gr_list
-                   if entry[1] != base_mt or
-                   (entry[0], entry[2], entry[3], entry[4], entry[5])
-                   not in n2_keys]
-
-        if debug:
-            print(f"Phase 1: {len(gr_list)} GR entries")
-            for i, (t, mt, ts, te, ks, ke, g) in enumerate(gr_list):
-                loads = ((te - ts) // g.mn) * ((ke - ks) // g.k)
-                print(f"  [{i}] {t:2s} {fmt_mt(mt)} tiles[{ts},{te - 1}] k[{ks},{ke - 1}] "
-                      f"gr_gran(mn={g.mn},k={g.k}) loads={loads}")
-
-        return gr_list
-
-    def _build_lr_conflict_map(self, schedule: LogicalSchedule):
-        """Build per-partition LR(MT n) info for LDS conflict checking.
-
-        Returns dict: (partition_idx, tensor) -> list of
-                      (subIterK_slot, k_start, k_end).
-        """
-        lr_mt_n_info = {}
-        for pi, partition_slots in enumerate(schedule):
-            for slot in partition_slots:
-                for lr in slot.lrs:
-                    if lr.mtIteration == 0:
-                        lr_mt_n_info.setdefault((pi, lr.tensor), []).append(
-                            (slot.subIterK,
-                             lr.tiles.subIterK_start,
-                             lr.tiles.subIterK_end))
-        return lr_mt_n_info
-
-    @staticmethod
-    def _has_lr_conflict(lr_mt_n_info, tensor, mt_val, pi, subIterK,
-                         gr_k_start, gr_k_end):
-        """Return True if placing GR(mt_val) at (pi, subIterK) conflicts.
-
-        GR(MT n+2) writes the same LDS buffer as MT n, so it conflicts
-        only if a later LR(MT n) in the same partition accesses an
-        overlapping subIterK range.
-        """
-        if mt_val != 2:
-            return False
-        for lr_slot, lr_ks, lr_ke in lr_mt_n_info.get((pi, tensor), []):
-            if lr_slot > subIterK and gr_k_start < lr_ke and lr_ks < gr_k_end:
-                return True
-        return False
-
-    def _distribute_grs(self, schedule: LogicalSchedule, gr_list, lr_mt_n_info, debug=False):
-        """Phase 2: Distribute GR atoms across partition × subIterK slots.
-
-        Explodes GR entries into atomic loads, distributes them into flat
-        buckets respecting LDS conflict constraints and load balance,
-        then remerges consecutive atoms and places them into partitions.
-        """
-        cfg = self.config
-        numK = cfg.numSubIterK
-        numP = cfg.numPartitions
-        numSlots = numP * numK
-
-        # 2a. Explode GR entries into atomic loads (1 load each)
-        atoms = []
-        for tensor, mt_val, t_start, t_end, k_start, k_end, gr_gran in gr_list:
-            mn = gr_gran.mn
-            for pos in range(t_start, t_end, mn):
-                atoms.append((tensor, mt_val, pos, pos + mn, k_start, k_end))
-
-        loads_per_slot = len(atoms) // numSlots
-
-        # 2b. Distribute atoms into flat buckets [0..numSlots),
-        #     each bucket maps to (partition=flat//numK, subIterK=flat%numK)
-        buckets = [[] for _ in range(numSlots)]
-        for atom in atoms:
-            tensor, mt_val, _, _, ks, ke = atom
-            cur = 0
-            while cur < numSlots - 1:
-                pi = cur // numK
-                subK = cur % numK
-                if (not self._has_lr_conflict(lr_mt_n_info, tensor, mt_val,
-                                              pi, subK, ks, ke) and
-                        len(buckets[cur]) < loads_per_slot):
-                    break
-                cur += 1
-            buckets[cur].append(atom)
-
-        if debug:
-            print(f"Phase 2b: {len(atoms)} atoms, {numSlots} slots, "
-                  f"{loads_per_slot} per slot")
-            for flat, bucket in enumerate(buckets):
-                pi = flat // numK
-                si = flat % numK
-                if bucket:
-                    items = ", ".join(
-                        f"{t} {fmt_mt(mt)} tile[{ts},{te-1}] k[{ks},{ke-1}]"
-                        for t, mt, ts, te, ks, ke in bucket)
-                    print(f"  P{pi} s{si}: {len(bucket)} atoms — {items}")
-                else:
-                    print(f"  P{pi} s{si}: empty")
-
-        # 2c. Remerge consecutive atoms and place into partitions
-        for flat, bucket in enumerate(buckets):
-            pi = flat // numK
-            si = flat % numK
-            target_slot = schedule[pi][si]
-            for atom in bucket:
-                tensor, mt_val, ts, te, ks, ke = atom
-                if target_slot.grs:
-                    prev = target_slot.grs[-1]
-                    if (prev.tensor == tensor and
-                            prev.mtIteration == mt_val and
-                            prev.tiles.subIterK_start == ks and
-                            prev.tiles.subIterK_end == ke and
-                            prev.tiles.tileId_end == ts):
-                        prev.tiles = MFMATileRange(ks, ke, prev.tiles.tileId_start, te)
-                        continue
-                target_slot.grs.append(GRPlacement(
-                    tensor=tensor, mtIteration=mt_val,
-                    tiles=MFMATileRange(ks, ke, ts, te),
-                    subIterK_slot=si,
-                    partition=pi))
-
-    def place_GRs(self, schedule: LogicalSchedule) -> LogicalSchedule:
-        """Place Global Reads by iterating MFMAs across partitions.
-
-        Phase 1: Build ordered GR list from partition traversal respecting gr granularities.
-        Phase 2: Distribute evenly GR atoms across all (partition, subIterK) slots. GR atoms being the smallest load granularity for a specific tensor.
-
-        This should give a sheduling respecting the following rules:
-         - GR are in the order we expect them from the LR pov
-         - we respect the GR granularities (can change the above rule a bit)
-         - Overall loads are spread accross all subIterKs of all partitions.
-
-        """
-        part_ranges = [self._partition_tile_range(pi)
-                       for pi in range(self.config.numPartitions)]
-
-        # TODO: cover PGR3 (offsetMT and offsetPartition may differ)
-        offsetMT = 1
-        offsetPartition = 1
-        # Build ordered list of GRs to place for the entire MT based on the partitioning ordering and the GR granularities.
-        gr_list = self._build_gr_list(schedule, part_ranges, offsetMT, offsetPartition)
-        # Map to keep track of LR(MT n) for each partiion and tensor, used for LDS double buffer conflict checking when placing GRs.
-        lr_mt_n_info = self._build_lr_conflict_map(schedule)
-        # Distribute GRs accross partition.
-        self._distribute_grs(schedule, gr_list, lr_mt_n_info)
-        return schedule
+        """Return {'A': (start, end), 'B': (start, end)} for partition pi."""
+        return partition_tile_range(self.config, pi)
 
     # ── Annotate dependencies ─────────────────────────────
 
@@ -1050,7 +217,7 @@ class LogicalScheduler:
             # Uses lr_by_tensor (all LRs across partitions) so that a more recent
             # LR loading a different subIterK still subsumes older data deps.
             if slot.mfma:
-                for t in self.tensors:
+                for t in self.config.tensors:
                     deps_for_t = []
                     for lr in lr_by_tensor.get(t, []):
                         if _tiles_overlap(slot.mfma, t, lr.tiles):
@@ -1106,7 +273,7 @@ class LogicalScheduler:
         def _dep_exec_order(dep):
             return (dep.mt_offset, dep.ref.partition, dep.ref.subIterK_slot)
 
-        for tensor in self.tensors:
+        for tensor in self.config.tensors:
             lr_with_gr_deps = []
             for pi, slots in enumerate(schedule):
                 for slot in slots:
@@ -1665,11 +832,14 @@ class LogicalScheduler:
             stop_after: If given, stop after the named pass and return early.
                 Used by tests to run the pipeline up to a specific stage.
         """
-        schedule: LogicalSchedule = self.place_LRs()
+        schedule: LogicalSchedule = place_lrs(self.config)
         if stop_after == 'place_LRs': return schedule
-        self.assign_vgpr_tiles(schedule)
+        alloc = assign_vgpr_tiles(self.config, schedule)
+        self.unroll_factor = alloc.unroll_factor
+        self.needs_unrolling = alloc.needs_unrolling
+        self.tile_peaks = alloc.tile_peaks
         if stop_after == 'assign_vgpr_tiles': return schedule
-        schedule = self.place_GRs(schedule)
+        schedule = place_grs(self.config, schedule)
         if stop_after == 'place_GRs': return schedule
         annotated: AnnotatedSchedule = self.annotate_deps(schedule)
         if stop_after == 'annotate_deps': return annotated
@@ -1792,7 +962,7 @@ class LogicalScheduler:
         return [GRPlacement(tensor=tensor, mtIteration=mt,
                             tiles=tiles['A' if tensor in ('A', 'SA') else 'B'],
                             subIterK_slot=0)
-                for tensor in self.tensors]
+                for tensor in self.config.tensors]
 
     def _preloop_make_lr(self, tiles: dict, schedule: AugmentedSchedule) -> List[LRPlacement]:
         """Create LR placements for first partition.
@@ -1805,7 +975,7 @@ class LogicalScheduler:
         first_mfma = schedule[0][0].mfma
 
         placements = []
-        for tensor in self.tensors:
+        for tensor in self.config.tensors:
             lr = LRPlacement(
                 tensor=tensor, mtIteration=0,
                 tiles=tiles[tensor],
@@ -1817,7 +987,7 @@ class LogicalScheduler:
 
     def _make_tensor_depops(self, cls) -> List[BaseOp]:
         """Create a BaseOp subclass instance for each tensor."""
-        return [cls(tensor=tensor) for tensor in self.tensors]
+        return [cls(tensor=tensor) for tensor in self.config.tensors]
 
     def build_preloop(self, schedule: AugmentedSchedule) -> EmittedSchedule:
         """Build preloop: pipeline initialization sequence before mainloop.
@@ -2189,7 +1359,7 @@ class LogicalScheduler:
                         m = slot.mfma
                         tiles_str = ""
                         parts = []
-                        for tensor in self.tensors:
+                        for tensor in self.config.tensors:
                             maps = m.vgpr_tile_maps.get(tensor)
                             if maps:
                                 parts.append(f"{tensor}:" + str(maps[ui]))
